Modifying .last to return a single element
Despite .last will always return one element, we weren't doing any
processing with the array returned by .get_last, so we were returning an
array with one element. We now modify the logic to get the first element (which will also be the only one) directly.

We also increase the Gem version to prepare a new release with this
change