Filter BG readings older than the specified minutes (#4)
The `Dexcom::BloodGlucose.get_last` method accepts a `minutes` argument, to specify that we want to get the BG readings published since those minutes ago. Currently, we are only using it to calculate how many datapoints we had to ask, knowing that Dexcom publishes a reading every 5 minutes. For example, if we call `Dexcom::BloodGlucose.get_last(minutes: 20)`, the code will get the last 4 published readings.

However, if the data  in Dexcom is outdated (e.g. because the transmitter lost connection and couldn't send the latest data), those 4 readings could be more than 20 minutes old. This is counterintuitive, because to get a fixed set of datapoints we already have the `max_count` argument.

This PR adds logic to filter out readings according to the timestamp whenever the minutes argument is used.

---
Commits: 

* Using a fixed time in tests

Introducing the Timecop gem to run tests at a fixed time

* Using helper methods to get mock Dexcom data

Adding two methods, one to get a single item at the specified datetime,
and another one to get an array of items when the last one is at the
specified datetime.

* Return different_timestamps in mock_api_bgs helper

Modify the mock_api_bgs helper function to return an array of values
separated by a 5-minute difference. This mimics the usual API behavior
and will make the data used in the tests closer to the real one

* Minor test refactor

Minor test refactor to avoid using variables directly on specs

* Clarifying tests with more variables and meaningful names

* Filtering BG readings out of the minutes range

Adding logic to remove datapoints out of the specified minutes range.
This will help when the data in Dexcom is outdated (e.g. because the
transmitter lost connection and couldn't send the latest datapoints)

* Cleaning the code + Adding ActiveSupport dependency

Refactoring the BloodGlucose::ClassMethods module to make the code more
readable.

Adding ActiveSupport library as a dependency to enable the
`.minutes.ago` and `.present?` methods